+++ src/cgalMesh.h
@@ -149,9 +149,9 @@ namespace PMP = CGAL::Polygon_mesh_processing;
 
 // -------------------------------------------------------------------------- //
 template <typename MeshT, typename PointT>
-MeshT csoup2mesh(std::vector<PointT>, std::vector<std::vector<int>>, const bool);
+MeshT csoup2mesh(std::vector<PointT>, std::vector<std::vector<size_t>>, const bool);
 
-std::vector<std::vector<int>> list_to_faces(const Rcpp::List);
+std::vector<std::vector<size_t>> list_to_faces(const Rcpp::List);
 
 template <typename PointT>
 std::vector<PointT> matrix_to_points3(const Rcpp::NumericMatrix);
@@ -168,7 +168,7 @@ Rcpp::List getFaces(MeshT&);
 
 void Message(std::string);
 
-EMesh3 readMeshFile(const std::string, bool);
+EMesh3 readMeshFile(const std::string, bool, bool);
 void writeMeshFile(
   const std::string, const int, const bool, std::string, EMesh3&
 );