+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - The `$clip` method with `volume=TRUE` didn't correctly preserve the face properties (e.g. the face colors). Now it partially works: the face properties of the clipped mesh are correctly preserved, but not the face properties of the clipping mesh (help wanted).
 
+- It is now possible to write a binary PLY file representing a mesh.
+
 
 # cgalMeshes 2.0.0